Split C match emission
